requirements.txt
flask>=3.0.0
moviepy>=2.0.0
Pillow>=10.0.0
requests>=2.31.0
anthropic>=0.40.0
openai-whisper>=20231117
numpy>=1.24.0
python-dotenv>=1.0.0
fal-client>=0.5.0
elevenlabs>=1.0.0
apscheduler>=3.10.0
